Separate out the recording processing
